agent/requirements.txt
# Pydantic AI with AG-UI support
pydantic-ai-slim[ag-ui]>=0.1.0

# CopilotKit Python SDK
copilotkit>=0.1.0

# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0

# Database
psycopg[binary]>=3.2.0

# Environment
python-dotenv>=1.0.0

# Google Gemini (or OpenAI)
pydantic-ai-slim[google]>=0.1.0
# openai>=1.50.0  # Uncomment if using OpenAI instead
